add run stats

+++ Assets/Scripts/Weapon/Deagle.cs
@@ -4,7 +4,7 @@ public class Deagle : Weapon
 {
     public override string Name => "Deagle";
     public override int ClipSize => 7;
-    public override float BaseDamage => 45f;
+    public override int BaseDamage => 45;
     public override float FireRate => 0.63f;
     public override float ReloadSpeed => 1.25f;
     public override float ReadySpeed => 0.35f;
