+++ Assets/Scripts/Weapon/RocketLauncher.cs
@@ -4,7 +4,7 @@ public class RocketLauncher : Weapon
 {
     public override string Name => "Rocket Launcher";
     public override int ClipSize => 7;
-    public override float BaseDamage => 65f;
+    public override int BaseDamage => 65;
     public override float FireRate => 0.70f;
     public override float ReloadSpeed => 1.25f;
     public override float ReadySpeed => 0.65f;
